V.1

+++ src/main/java/com/lucas/imobiliaria/model/domain/repository/UsuariosRepository.java
@@ -4,6 +4,6 @@ import com.lucas.imobiliaria.model.domain.users.Usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 
-public interface UsuariosRepository extends JpaRepository<Usuarios, Integer> {
+public interface UsuariosRepository extends JpaRepository<Usuarios, Long> {
     UserDetails findByEmail(String email);
 }
